requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.2
greenlet==3.0.2
huggingface-hub==0.19.4
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langsmith==0.0.70
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
